Change the use of the :LOCAL-INTERNET-DOMAINS site variable.
If a host is specified (e.g. by an unknown-host function) with one of
these domains, we try to make the shortest prefix (w/o the domain) a
short-name for the host.  And if trying to find a host name without
dots in it, search the :LOCAL-INTERNET-DOMAINS for it.

We do NOT just ignore/strip the domain when parsing a host name.  (The
other behaviour described for the site variable seems never to have
been implemented.)


network/chaos/chuse.lisp (FIND-SHORTEST-LOCAL-DOMAIN-PREFIX)
(CHAOS-UNKNOWN-HOST-FUNCTION-1): New functions.
(CHAOS-UNKNOWN-HOST-FUNCTION): Call CHAOS-UNKNOWN-HOST-FUNCTION-1
depending on if we are looking up a local domain host, or not.
network/host.lisp (LOCAL-INTERNET-DOMAINS): Update documentation.
(PARSE-HOST): Don't try to look up local-internet-domain hosts, let
CHAOS:CHAOS-UNKNOWN-HOST-FUNCTION do the job.